+++ R/emmGrid-methods.R
@@ -198,11 +198,6 @@ vcov.emmGrid = function(object, ...) {
 #' \item{\code{adjust}}{(\code{character)}) is the default for the \code{adjust}
 #' argument in \code{\link{summary.emmGrid}}.}
 #' 
-#' \item{\code{estType}}{(\code{character}) is the type of the estimate. It
-#' should match one of \samp{c("prediction", "contrast", "pairs")}. This is used
-#' along with \code{"adjust"} to determine appropriate adjustments to P values
-#' and confidence intervals.}
-#' 
 #' \item{\code{famSize}}{(integer) is the number of means involved in a family of
 #' inferences; used in Tukey adjustment}
 #' 
@@ -233,6 +228,16 @@ vcov.emmGrid = function(object, ...) {
 #' \code{\link{predict.emmGrid}}, and \code{\link{emmip}}. Valid values are
 #' \code{"link"} (with synonyms \code{"lp"} and \code{"linear"}), or
 #' \code{"response"}.}
+#' 
+#' \item{\code{bias.adjust}, \code{frequentist}}{(character) These
+#' are used by \code{summary} if the value of these arguments are not specified.}
+#' 
+#' \item{\code{estType}}{(\code{character}) is used internally to determine 
+#' what \code{adjust} methods are appropriate. It should match one of 
+#' \samp{c("prediction", "contrast", "pairs")}. As an example of why this is needed,
+#' the Tukey adjustment should only be used for pairwise comparisons 
+#' (\code{estType = "pairs"}); if \code{estType} is some other string, Tukey
+#' adjustments are not allowed.}
 #' 
 #' \item{\code{avgd.over}}{(\code{character)} vector) are the names of the 
 #' variables whose levels are averaged over in obtaining marginal averages of 
@@ -284,12 +289,9 @@ vcov.emmGrid = function(object, ...) {
 #' emmeans(mypigs.rg, "source")
 update.emmGrid = function(object, ..., silent = FALSE) {
     args = list(...)
-    valid.misc = c("adjust","alpha","avgd.over","by.vars","delta","df",
-                   "initMesg","estName","estType","famSize","infer","inv.lbl",
-                   "level","methDesc","nesting","null","predict.type","pri.vars"
-                   ,"side","sigma","tran","tran.mult","tran.offset","tran2","type","is.new.rg")
+    # see .valid.misc below this function for list of legal options
     valid.slots = slotNames(object)
-    valid.choices = union(valid.misc, valid.slots)
+    valid.choices = union(.valid.misc, valid.slots)
     misc = object@misc
     for (nm in names(args)) {
         fullname = try(match.arg(nm, valid.choices), silent=TRUE)
@@ -348,6 +350,13 @@ update.emmGrid = function(object, ..., silent = FALSE) {
     object@misc = misc
     object
 }
+
+### List of valid strings to match in update() ###
+.valid.misc = c("adjust","alpha","avgd.over","bias.adjust","by.vars","delta","df",
+               "initMesg","estName","estType","famSize","frequentist","infer","inv.lbl",
+               "level","methDesc","nesting","null","predict.type","pri.vars"
+               ,"side","sigma","tran","tran.mult","tran.offset","tran2","type","is.new.rg")
+
 
 #' Set or change emmeans options
 #'